requirements.txt
python-dotenv
psycopg2-binary
langchain-openai
langchain-core
ragflow-sdk
fastapi
uvicorn
fastmcp
